+++ heat/core/communication.py
@@ -170,21 +170,21 @@ class MPICommunication(Communication):
         Parameters
         ----------
         shape : Tuple[int,...]
-            The global shape of the data to be split
+            The global shape of the data to be split.
         split : int
-            The axis along which to chunk the data
+            The axis along which to chunk the data. Must be within the range of ``shape``.
         rank : int, optional
             Process for which the chunking is calculated for, defaults to ``self.rank``.
-            Intended for creating chunk maps without communication
+            Intended for creating chunk maps without communication.
         w_size : int, optional
             The MPI world size, defaults to ``self.size``.
-            Intended for creating chunk maps without communication
-
+            Intended for creating chunk maps without communication.
         """
-        # ensure the split axis is valid, we actually do not need it
-        split = sanitize_axis(shape, split)
         if split is None:
             return 0, shape, tuple(slice(0, end) for end in shape)
+        if split < 0:
+            split = len(shape) + split
+
         rank = self.rank if rank is None else rank
         w_size = self.size if w_size is None else w_size
         if not isinstance(rank, int) or not isinstance(w_size, int):
@@ -435,6 +435,25 @@ class MPICommunication(Communication):
         recvtypes = [mpi_type] * nproc
 
         return self.as_mpi_memory(obj), (recvcount, recvdispls), recvtypes
+
+    def Free(self) -> None:
+        """
+        Free a communicator.
+        """
+        self.handle.Free()
+
+    def Split(self, color: int = 0, key: int = 0) -> MPICommunication:
+        """
+        Split communicator by color and key.
+
+        Parameters
+        ----------
+        color : int, optional
+            Determines the new communicator for a process.
+        key: int, optional
+            Ordering within the new communicator.
+        """
+        return MPICommunication(self.handle.Split(color, key))
 
     def Irecv(
         self,